fix model album

+++ controllers/album.js
@@ -18,27 +18,19 @@ const getAlbum = (req, res, next) => {
 const deleteAlbum = (req, res, next) => {
   const { id: _id } = req.params;
   console.log('delete album');
+  if (!req.user) {
+    return next(new NotAuthError(notAuthErrors.noAuth));
+  }
 
-  Album.findByIdAndDelete({ _id })
+  return Album.findByIdAndDelete({ _id })
     .then((album) => {
       const { createdAt } = album;
       const folderNameAlbum = createdAt.toISOString().slice(0, 16).replace(':', '');
       const dirPath = path.join('/home/newrdlink/projects/vs/backend/public/album', folderNameAlbum);
 
       if (fs.existsSync(dirPath)) {
-        console.log(1, 'folder founded');
-        // // for remove dir from localhost DB and location file
-        // const pathFileName = path.normalize(cutExpStr(worker.image));
-        // fs.rmdirSync(preparePathForRmDir(pathFileName), { recursive: true });
-        // for remove dir from serverDB location file
-        // const dirPath = path.join('/home/newrdlink/projects/
-        // vs/backend/public/', preparePathForRmDir(worker.image));
+        console.log('folder founded');
         fs.rmdirSync(dirPath, { recursive: true });
-        // console.log(2, dirPath);
-        // News.findByIdAndRemove(_id)
-        //   .then(() => console.log('новость удалена из базы'))
-        //   .catch(next);
-        // res.send(news);
       }
       res.send(album);
     })
@@ -46,16 +38,13 @@ const deleteAlbum = (req, res, next) => {
 };
 
 const createAlbum = (req, res, next) => {
-  // console.log('create album');
   if (!req.user) {
-    // console.log('error');
     return next(new NotAuthError(notAuthErrors.noAuth));
   }
   const albumData = JSON.parse(req.body.albumData);
-  // console.log(req.body);
+
   const imagesFront = req.files.imageFilesAlbum;
-  // console.log(1, albumData);
-  // console.log(2, imagesFront);
+
   const images = [];
   const folderNameAlbum = albumData.createdAt.slice(0, 16).replace(':', '');
   const dirPath = path.join(__dirname, '..', 'public/album', folderNameAlbum);
